+++ ios/CHANGELOG.md
@@ -23,6 +23,9 @@ Line wrap the file at 100 chars.                                              Th
 
 
 ## [Unreleased]
+### Added
+- Add option to block gambling and adult content.
+
 ### Fixed
 - Improve random port distribution. Should be less biased towards port 53.